requirements.txt
asgiref==3.8.1
Django==5.0.7
djangorestframework==3.15.2
Faker==26.0.0
python-dateutil==2.9.0.post0
python-decouple==3.8
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.2
